% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allowedCatchCat}
\alias{allowedCatchCat}
\title{Allowed CatchCat variable names}
\format{
A character vector
}
\source{
WGRDBES-EST
}
\usage{
allowedCatchCat
}
\description{
A vector containing the allowed names
}
\keyword{datasets}
