% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allowedEffortVariable}
\alias{allowedEffortVariable}
\title{Allowed effortVariable names}
\format{
A character vector
}
\source{
WGRDBES-EST
}
\usage{
allowedEffortVariable
}
\description{
A vector containing the allowed effortVariable names
}
\keyword{datasets}
