% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allowedLandingsVariable}
\alias{allowedLandingsVariable}
\title{Allowed landingsVariable names}
\format{
A character vector
}
\source{
WGRDBES-EST
}
\usage{
allowedLandingsVariable
}
\description{
A vector containing the allowed landingVariable names
}
\keyword{datasets}
