% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allowedSamplingVariable}
\alias{allowedSamplingVariable}
\title{Allowed samplingVariable names}
\format{
A character vector
}
\source{
WGRDBES-EST
}
\usage{
allowedSamplingVariable
}
\description{
A vector containing the allowed samplingVariable names
}
\keyword{datasets}
