% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{barPlotsByGroupingVariable}
\alias{barPlotsByGroupingVariable}
\title{Internal function to return a list of barplots which compare the
groupingVariable in landings, effort, and sample data.}
\usage{
barPlotsByGroupingVariable(
  landingsData = NA,
  effortData = NA,
  sampleData = NA,
  vesselFlag,
  catchCat,
  quarter,
  landingsVariable,
  effortVariable,
  samplingVariable,
  groupingVariable,
  topN = NA,
  plotQuarters = FALSE
)
}
\arguments{
\item{landingsData}{Landings data}

\item{effortData}{Effort data}

\item{sampleData}{Sample data}

\item{vesselFlag}{Registered Country of Vessel - e.g "IE", "ES" or "FR".}

\item{catchCat}{Catch category}

\item{quarter}{Quarter of year}

\item{landingsVariable}{Variable from CE to plot}

\item{samplingVariable}{variable from SA to plot}
}
\value{
A tagList of plotly plots
}
\description{
Internal function to return a list of barplots which compare the
groupingVariable in landings, effort, and sample data.
}
