% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageSpatialBivariate.R
\name{bivariatePlot}
\alias{bivariatePlot}
\title{Internal function to return a list of plots which compare the statistical
rectangle where landings occured to the statistical rectangles where
sampling occured.  The relative amounts of landings/samples are
shown by the use of different colours for the rectangle.}
\usage{
bivariatePlot(
  landingsData = NA,
  effortData = NA,
  sampleData,
  vesselFlag,
  catchCat,
  landingsVariable,
  effortVariable,
  samplingVariable
)
}
\arguments{
\item{landingsData}{Landings data}

\item{effortData}{Effort data}

\item{sampleData}{Sample data}

\item{vesselFlag}{Registered Country of Vessel - e.g "IE", "ES" or "FR".}

\item{catchCat}{Catch category}

\item{landingsVariable}{The variable from the landings data to plot}

\item{effortVariable}{The variable from the effort data to plot}

\item{samplingVariable}{The variable from the sample data to plot}
}
\value{
A tagList of ggplot2 plots
}
\description{
Internal function to return a list of plots which compare the statistical
rectangle where landings occured to the statistical rectangles where
sampling occured.  The relative amounts of landings/samples are
shown by the use of different colours for the rectangle.
}
