% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bivariatePlotCreateGraph}
\alias{bivariatePlotCreateGraph}
\title{Internal function to create bivariate plot}
\usage{
bivariatePlotCreateGraph(dataToPlot, ices_rect, title, subtitle, xlab, ylab)
}
\arguments{
\item{dataToPlot}{The data to plot}

\item{ices_rect}{ICES rectangles spatial information}

\item{title}{Title of the graph to create}

\item{subtitle}{Subtitle of the graph to create}

\item{xlab}{x-axis label for the graph to create}

\item{ylab}{y-axis label of the graph to create}
}
\value{
ggplot graph
}
\description{
Internal function to create bivariate plot
}
