% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageBySpecies.R
\name{coverageBySpecies}
\alias{coverageBySpecies}
\title{Provides graphical outputs to compare the species in
landings to those in the sample data.}
\usage{
coverageBySpecies(
  dataToPlot,
  year = NA,
  vesselFlag = NA,
  landingsVariable = c("CLoffWeight", "CLsciWeight"),
  samplingVariable = c("SAsampWtLive", "SAnumSamp", "SAsampWtMes"),
  catchCat = c("Lan", "Dis", "Catch"),
  includeLandings = TRUE,
  includeSamples = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{dataToPlot}{RDBES data to be plotted (as an RDBESDataObject)}

\item{year}{Year to be assessed e.g 2021}

\item{vesselFlag}{Registered Country of Vessel - e.g "IE", "ES" or "FR".}

\item{landingsVariable}{Landings variable to be assessed}

\item{samplingVariable}{Sampling Variable to be assessed}

\item{catchCat}{Sampling catch category - "Lan", "Dis",or "Catch"}

\item{includeLandings}{(Optional) Set to TRUE to include landings in the
plot - default is TRUE.}

\item{includeSamples}{(Optional) Set to TRUE to include samples in the
plot - default is TRUE.}

\item{verbose}{(Optional) Set to TRUE to print more information. Default is
FALSE}
}
\value{
A tagList of plotly plots
}
\description{
Provides graphical outputs to compare the species in
landings to those in the sample data.
}
\examples{
\dontrun{

myH1RawObject <- RDBEScore::createRDBESDataObject(
  rdbesExtractPath = "./tests/testthat/h1_v_1_19_13"
)

myYear <- 1965
myvesselFlag <- "ZW"

myPlots <- coverageLandingsByGear(
  dataToPlot = myH1RawObject,
  year = myYear,
  vesselFlag = myvesselFlag,
  catchCat = "Lan"
)

myPlots[1]
}
}
