% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageLandings.R
\name{coverageLandings}
\alias{coverageLandings}
\title{Provides graphical outputs to compare sampled data against landings
using the RDBES format.}
\usage{
coverageLandings(
  dataToPlot,
  year = NA,
  quarter = NA,
  vesselFlag = NA,
  var = c("species", "gear", "Statrec"),
  commercialVariable = c("CLoffWeight", "CLsciWeight"),
  samplingVariable = c("SAsampWtLive", "SAnumSamp", "SAsampWtMes"),
  catchCat = c("Lan", "Dis", "Catch"),
  spatialPlot = c("Bivariate", "Points"),
  verbose = FALSE
)
}
\arguments{
\item{dataToPlot}{RDBES data to be plotted (as an RDBESDataObject)}

\item{year}{Year to be assessed e.g 2021}

\item{quarter}{Quarter to be assessed - possible choices 1,2,3 or 4.}

\item{vesselFlag}{Registered Country of Vessel - e.g "IE", "ES" or "FR".}

\item{var}{Variable of interest - "gear" or "Statrec"}

\item{commercialVariable}{Landings variable to be assessed}

\item{samplingVariable}{Sampling Variable to be assessed}

\item{catchCat}{Sampling catch category - landings, catch or discards}

\item{spatialPlot}{Type of Spatial plot to return - bivariate or points}

\item{verbose}{(Optional) Set to TRUE to print more information. Default is
FALSE}
}
\value{
A tagList of plots
}
\description{
Provides graphical outputs to compare sampled data against landings
using the RDBES format.
}
\examples{
\dontrun{

myH1RawObject <-
  createRDBESDataObject(rdbesExtractPath = "tests/testthat/h1_v_1_19_13")
myH1EstObj <- createRDBESEstObject(myH1RawObject, 1)

# Example 1:
coverageLandings(
  dataToPlot = myH1EstObj,
  year = 2018,
  vesselFlag = "FR",
  var = "species",
  catchCat = "Lan"
)

}
}
