% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageLandingsByGear.R
\name{coverageLandingsByGear}
\alias{coverageLandingsByGear}
\title{Provides graphical outputs to compare the fishing gears
in landings to those in the sample data.}
\usage{
coverageLandingsByGear(
  dataToPlot,
  year = NA,
  quarter = NA,
  vesselFlag = NA,
  catchCat = c("Lan", "Dis", "Catch"),
  verbose = FALSE
)
}
\arguments{
\item{dataToPlot}{RDBES data to be plotted (as an RDBESDataObject)}

\item{year}{Year to be assessed e.g 2021}

\item{quarter}{Quarter to be assessed - possible choices 1,2,3 or 4.}

\item{vesselFlag}{Registered Country of Vessel - e.g "IE", "ES" or "FR".}

\item{catchCat}{Sampling catch category - "Lan", "Dis",or "Catch"}

\item{verbose}{(Optional) Set to TRUE to print more information. Default is
FALSE}
}
\value{
A tagList of plotly plots
}
\description{
Provides graphical outputs to compare the fishing gears
in landings to those in the sample data.
}
\examples{
\dontrun{

  myH1RawObject <- RDBEScore::createRDBESDataObject(
                  rdbesExtractPath = "../tests/testthat/h1_v_1_19_13")

  myYear = 1965
  myvesselFlag = "ZW"

  myPlots <- coverageLandingsByGear(
    dataToPlot = myH1RawObject,
    year = myYear,
    vesselFlag = myvesselFlag,
    catchCat = "Lan"
    )

  myPlots[1]

}
}
