% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageLandingsBySpecies.R
\name{coverageLandingsBySpecies}
\alias{coverageLandingsBySpecies}
\title{Provides graphical outputs to compare the species in
landings to those in the sample data.}
\usage{
coverageLandingsBySpecies(
  dataToPlot,
  year = NA,
  vesselFlag = NA,
  catchCat = c("Lan", "Dis", "Catch"),
  verbose = FALSE
)
}
\arguments{
\item{dataToPlot}{RDBES data to be plotted (as an RDBESDataObject)}

\item{year}{Year to be assessed e.g 2021}

\item{vesselFlag}{Registered Country of Vessel - e.g "IE", "ES" or "FR".}

\item{catchCat}{Sampling catch category - "Lan", "Dis",or "Catch"}

\item{verbose}{(Optional) Set to TRUE to print more information. Default is
FALSE}
}
\value{
A tagList of plotly plots
}
\description{
Provides graphical outputs to compare the species in
landings to those in the sample data.
}
\examples{
\dontrun{

  myH1RawObject <- RDBEScore::createRDBESDataObject(
                  rdbesExtractPath = "./tests/testthat/h1_v_1_19_13")

  myYear = 1965
  myvesselFlag = "ZW"

  coverageLandingsByGear(
    dataToPlot = myH1RawObject,
    year = myYear,
    vesselFlag = myvesselFlag,
    catchCat = "Lan"
  )
}
}
