% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageLandingsSpatialBivariate.R
\name{coverageLandingsSpatialBivariate}
\alias{coverageLandingsSpatialBivariate}
\title{Provides graphical outputs which compare the statistical
rectangle where landings occured to the statistical rectangles where
sampling occured.  The relative amounts of landings/samples are
shown by the use of different colours for the rectangle.}
\usage{
coverageLandingsSpatialBivariate(
  dataToPlot,
  year = NA,
  vesselFlag = NA,
  commercialVariable = c("CLoffWeight", "CLsciWeight"),
  samplingVariable = c("SAsampWtLive", "SAnumSamp", "SAsampWtMes"),
  catchCat = c("Lan", "Dis", "Catch"),
  verbose = FALSE
)
}
\arguments{
\item{dataToPlot}{RDBES data to be plotted (as an RDBESDataObject)}

\item{year}{Year to be assessed e.g 2021}

\item{vesselFlag}{Registered Country of Vessel - e.g "IE", "ES" or "FR".}

\item{commercialVariable}{Landings variable to be assessed}

\item{samplingVariable}{Sampling Variable to be assessed}

\item{catchCat}{Sampling catch category - "Lan", "Dis",or "Catch"}

\item{verbose}{(Optional) Set to TRUE to print more information. Default is
FALSE}
}
\value{
A tagList of ggplot2 plots
}
\description{
Provides graphical outputs which compare the statistical
rectangle where landings occured to the statistical rectangles where
sampling occured.  The relative amounts of landings/samples are
shown by the use of different colours for the rectangle.
}
\examples{
\dontrun{

  myH1RawObject <- RDBEScore::createRDBESDataObject(
                  rdbesExtractPath = "../tests/testthat/h1_v_1_19_13")

  myYear = 1965
  myvesselFlag = "ZW"

  coverageLandingsSpatialBivariate(
    dataToPlot = myH1RawObject,
    year = myYear,
    vesselFlag = myvesselFlag,
    catchCat = "Lan",
    commercialVariable = "CLoffWeight",
    samplingVariable = "SAsampWtLive"
    )

}
}
