% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createBarPlot}
\alias{createBarPlot}
\title{Internal function to create a bar plot}
\usage{
createBarPlot(
  dataToPlot,
  title,
  plotQuarters = FALSE,
  showLegend = FALSE,
  paletteToUse = "paletteForPlotting"
)
}
\arguments{
\item{dataToPlot}{The data to plot}

\item{title}{Title to use for the plot}

\item{plotQuarters}{Show bars coloured by quarter?}

\item{showLegend}{Show the legend?}
}
\value{
A plotly plot
}
\description{
Internal function to create a bar plot
}
