% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterLandingsDataForCoverage}
\alias{filterLandingsDataForCoverage}
\title{Internal function to filter landings data for coverageXXX functions}
\usage{
filterLandingsDataForCoverage(
  landingsData,
  yearToFilter = NA,
  quarterToFilter = NA,
  vesselFlag = NA,
  verbose = T
)
}
\arguments{
\item{landingsData}{A data table of landings data ("CL")}

\item{vesselFlag}{(Optional) Registered Country of Vessel - e.g "IE", "ES" or "FR"}

\item{verbose}{(Optional) Set to TRUE to print more information. Default is
FALSE}

\item{year}{(Optional) Year to be assessed e.g 2021}

\item{quarter}{(Optional) Quarter to be assessed - possible choices 1,2,3 or 4}
}
\value{
A data table of landings data ("CL")
}
\description{
Internal function to filter landings data for coverageXXX functions
}
