% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterSampleDataForCoverage}
\alias{filterSampleDataForCoverage}
\title{Internal function to filter sample data for coverageXXX functions}
\usage{
filterSampleDataForCoverage(
  sampleData,
  yearToFilter = NA,
  quarterToFilter = NA,
  vesselFlag = NA,
  catchCat = NA,
  verbose = T
)
}
\arguments{
\item{sampleData}{An RDBESEstObject}

\item{vesselFlag}{(Optional) Registered Country of Vessel - e.g "IE", "ES" or "FR".}

\item{catchCat}{(Optional) Sampling catch category - landings, catch or discards}

\item{verbose}{(Optional) Set to TRUE to print more information. Default is
FALSE}

\item{year}{(Optional) Year to be assessed e.g 2021}

\item{quarter}{(Optional) Quarter to be assessed - possible choices 1,2,3 or 4}
}
\value{
An RDBESEstObject
}
\description{
Internal function to filter sample data for coverageXXX functions
}
