% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatDataForBarPlotsByGroupingVariable}
\alias{formatDataForBarPlotsByGroupingVariable}
\title{Internal function to fromat the data ready for the createBarPlot()
function.}
\usage{
formatDataForBarPlotsByGroupingVariable(dataToFormat, topN)
}
\arguments{
\item{dataToFormat}{The data to plot}

\item{topN}{A value to limit the data plotted to the top N values}
}
\value{
A data frame ready for createBarPlot()
}
\description{
Internal function to fromat the data ready for the createBarPlot()
function.
}
