% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gearPlot}
\alias{gearPlot}
\title{Internal function to return a list of plots which compare the fishing gears
in landings to those in the sample data.}
\usage{
gearPlot(landingsData, sampleData, vesselFlag, catchCat, quarter)
}
\arguments{
\item{landingsData}{Landings data}

\item{sampleData}{Sample data}

\item{vesselFlag}{Registered Country of Vessel - e.g "IE", "ES" or "FR".}

\item{catchCat}{Catch category}

\item{quarter}{Quarter of year}
}
\value{
A tagList of plotly plots
}
\description{
Internal function to return a list of plots which compare the fishing gears
in landings to those in the sample data.
}
