% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageSpatial.R
\name{getQuantiles}
\alias{getQuantiles}
\title{Internal function to get quantiles - we'll use this for the scale when
plotting the spatial graph}
\usage{
getQuantiles(dataToProcess, numberOfClasses = 6)
}
\arguments{
\item{dataToProcess}{The data to process}

\item{numberOfClasses}{The number of classes we want in the quantiles}
}
\value{
The data with mean_quantiles appended
}
\description{
Internal function to get quantiles - we'll use this for the scale when
plotting the spatial graph
}
