% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icesAreaSF}
\alias{icesAreaSF}
\title{ICES statistical areas}
\format{
A data frame containing spatial information
\describe{
  \item{Area_Full}{Name of the ICES statistical area}...
}
}
\source{
ICES: https://gis.ices.dk/geonetwork/srv/api/records/c784a0a3-752f-4b50-b02f-f225f6c815eb
}
\usage{
icesAreaSF
}
\description{
An sf object containing spatial information for ICES statistical areas
}
\keyword{datasets}
