% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icesRects}
\alias{icesRects}
\title{ICES statistical rectangles}
\format{
A data frame containing spatial information
\describe{
  \item{ICESNAME}{Name of the stat rectangle}...
}
}
\source{
ICES
}
\usage{
icesRects
}
\description{
A data frame containing spatial information for ICES statistical rectangles
}
\keyword{datasets}
