% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icesRectSpatialPolygon}
\alias{icesRectSpatialPolygon}
\title{Shoreline spatial data}
\format{
A SpatialPolygonsDataFrame
}
\source{
ICES
}
\usage{
icesRectSpatialPolygon
}
\description{
A SpatialPolygonsDataFrame data frame containing spatial information for
ICES statistical rectangles
}
\keyword{datasets}
