% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageLandings.R
\name{pointsPlot}
\alias{pointsPlot}
\title{Internal function to return a list of plots which compare the statistical
rectangle where landings occured to the statitical rectangles where
sampling occured}
\usage{
pointsPlot(
  landingsData,
  sampleData,
  flagLabel,
  catchCat,
  commercialVariable,
  samplingVariable
)
}
\arguments{
\item{landingsData}{Landings data}

\item{sampleData}{Sample data}

\item{flagLabel}{Text to use for vessel flag description}

\item{catchCat}{Catch category}

\item{commercialVariable}{The variable from the landings data to plot}

\item{samplingVariable}{The variable from the sample data to plot}
}
\value{
A tagList of ggplot2 plots
}
\description{
Internal function to return a list of plots which compare the statistical
rectangle where landings occured to the statitical rectangles where
sampling occured
}
