% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocessEffortDataForCoverage}
\alias{preprocessEffortDataForCoverage}
\title{Internal function to prepare effort data for coverageXXX functions}
\usage{
preprocessEffortDataForCoverage(dataToPlot, verbose)
}
\arguments{
\item{dataToPlot}{An RDBESDataObject}

\item{verbose}{(Optional) Set to TRUE to print more information. Default is
FALSE}
}
\value{
A data table of effort data ("CE")
}
\description{
Internal function to prepare effort data for coverageXXX functions
}
