% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocessLandingsDataForCoverage}
\alias{preprocessLandingsDataForCoverage}
\title{Internal function to prepare landings data for coverageLandingsXXX functions}
\usage{
preprocessLandingsDataForCoverage(dataToPlot, verbose)
}
\arguments{
\item{dataToPlot}{An RDBESDataObject}

\item{verbose}{(Optional) Set to TRUE to print more information. Default is
FALSE}
}
\value{
A data table of landings data ("CL")
}
\description{
Internal function to prepare landings data for coverageLandingsXXX functions
}
