% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocessSampleDataForCoverage}
\alias{preprocessSampleDataForCoverage}
\title{Internal function to prepare sample data for coveragesXXX functions
NOT WORKING, to be check.}
\usage{
preprocessSampleDataForCoverage(dataToPlot, verbose)
}
\arguments{
\item{dataToPlot}{An RDBESDataObject}

\item{(Optional)}{Set to TRUE to print more information. Default is
FALSE}
}
\value{
an RDBESEstObject
}
\description{
Internal function to prepare sample data for coveragesXXX functions
NOT WORKING, to be check.
}
