% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessSamplingDataForCoverage.R, R/utils.R
\name{preprocessSamplingDataForCoverage}
\alias{preprocessSamplingDataForCoverage}
\title{Function perprocessing the Sampling data for coverage. Function based on
the function with the same name in utils.R. This function extended of data
validation and add variables from Biological Variables table, Frequency
Measure Table and general information occure in DE and SD ex. SamplingCountry,
hierarchy.The function works only for RDBESDataObject of one hierarchy.}
\usage{
preprocessSamplingDataForCoverage(dataToPlot, verbose)

preprocessSamplingDataForCoverage(dataToPlot, verbose)
}
\arguments{
\item{verbose}{Set to TRUE to print more information. Default is
FALSE}

\item{RDBESDataObject}{An RDBESDataObject.}

\item{validate}{Set to TRUE if you want validation to be carried out. The
default if TRUE.}

\item{strict}{(Optional) This function can validate its input data - should
the validation be strict? The default is TRUE.}

\item{bioVar}{set TRUE to add biological variable ex. Length, Weight, Age,
Sex, Maturity to the sample data. The default is FALSE.}

\item{generalVar}{set TRUE to add information about SamplingCountry, Hierarchy
and Sampling scheme (ex.Baltic SPF regional) (DE, SD table). The default is FALSE.}

\item{RDBESobj}{An RDBESDataObject}
}
\value{
Function returns table.

an RDBESEstObject
}
\description{
Function perprocessing the Sampling data for coverage. Function based on
the function with the same name in utils.R. This function extended of data
validation and add variables from Biological Variables table, Frequency
Measure Table and general information occure in DE and SD ex. SamplingCountry,
hierarchy.The function works only for RDBESDataObject of one hierarchy.

Internal function to prepare sample data for coveragesXXX functions
}
