% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produceSamplingCoverage.R
\name{produceSamplingCoverage}
\alias{produceSamplingCoverage}
\title{Provides tabular or graphical outputs which compare sampling versus 
landing coverage across the level of a factor variable specified by the user.}
\usage{
produceSamplingCoverage(
  RDBESobj,
  var = c("SAsampWtLive", "SAnumSamp", "SAamountSamp", "SAsampWtMes"),
  contrastVar = c("CLoffWeight", "CLsciWeight", "CLtotalOfficialLandingsValue",
    "CEnumFracTrips", "CEnumDomTrip"),
  by = NA,
  type = c("Spatial", "Time", "Fleet", "Species"),
  resolution = c("ICES Subdivision", "ICES Rectangle", "Year", "Semester", "Quarter",
    "Month", "Scale of fishery", "Species unit", "Species group"),
  year = NA,
  quarter = NA,
  vesselFlag = NA,
  catchCat = NA,
  showSamples = T,
  output_type = NA,
  verbose = T
)
}
\arguments{
\item{RDBESobj}{RDBES data to be plotted (as an RDBESDataObject)}

\item{var}{The unit of the variable of interest. Possible values are "WeightSampled", "NumberSamples", "IndividualsSampled".}

\item{contrastVar}{A variable to be selected by the user.  landing or effort selected by the user.}

\item{by}{(Optional) One (or more) factor(s) to split the data.}

\item{type}{Type of comparison to be represented by the plot. Possible values are "Spatial", "Time", "Fleet", "Species".}

\item{resolution}{The resolution at which the type is required. Possible values: "ICES Subdivision", "ICES Rectangle", "Year", "Semester", "Quarter", "Month", "ScaleFishery", "SpeciesList", "SpeciesGroupList".}

\item{year}{(Optional) Year to be assessed. Default is all years contained in RDBESobj data.}

\item{quarter}{(Optional) Quarter to be assessed - possible choices 1,2,3 or 4}

\item{vesselFlag}{(Optional) Registered Country of Vessel - e.g "IE", "ES" or "FR"}

\item{showSamples}{(Optional) Determines whether the samples data should be displayed or not. Default is TRUE.}

\item{output_type}{(Optional) Determines whether the output is tabular or a graphical plot. Possible values are "table" or "plot". Default is "plot".}

\item{verbose}{(Optional) Set to TRUE to print more information. Default is FALSE}
}
\value{
table or plot depending on user selection.
}
\description{
Provides tabular or graphical outputs which compare sampling versus 
landing coverage across the level of a factor variable specified by the user.
}
\examples{
\dontrun{


}
}
