% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDBESvisualise-package.R
\docType{package}
\name{RDBESvisualise-package}
\alias{RDBESvisualise}
\alias{RDBESvisualise-package}
\title{RDBESvisualise: Visualisation and Summarisation Tools for RDBES Data}
\description{
Provides utility functions to summarise and visualise data from Regional Database and Estimation System (RDBES).
}
\author{
\strong{Maintainer}: Colin Millar \email{colin.millar@ices.dk}

Authors:
\itemize{
  \item David Currie (\href{https://orcid.org/0000-0002-3523-6895}{ORCID})
  \item Richard Meitern \email{richard.meitern@ut.ee} (\href{https://orcid.org/0000-0002-2600-3002}{ORCID})
  \item Nuno Prista \email{nuno.prista@slu.se} (\href{https://orcid.org/0000-0002-5145-7241}{ORCID})
  \item Nicholas Carey \email{nicholas.carey@gov.scot}
  \item Petri Sarvamaa \email{Petri.Sarvamaa@luke.fi}
  \item Kirsten Birch Håkansson \email{kih@aqua.dtu.dk}
  \item Karolina Molla Gazi \email{karolina.mollagazi@wur.nl}
  \item Julia Wischnewski \email{julia.wischnewski@thuenen.de}
  \item Ana Cláudia Fernandes \email{acfernandes@ipma.pt} (\href{https://orcid.org/0000-0001-6041-8140}{ORCID})
  \item Katarzyna Krakówka \email{kkrakowka@mir.gdynia.pl}
  \item Marta Szymańska \email{msuska@mir.gdynia.pl}
  \item Nicolas Goñi \email{nicolas.goni@luke.fi}
}

Other contributors:
\itemize{
  \item Annica de Groote \email{annica.isaksson.de.groote@slu.se} [contributor]
  \item Jonathan Ball \email{johnathan.ball@cefas.co.uk} [contributor]
  \item Jonathan Rault \email{Jonathan.Rault@ifremer.fr} [contributor]
  \item Antti Sykkö \email{antti.sykko@luke.fi} [contributor]
  \item Liz Clarke \email{Liz.Clarke@gov.scot} [contributor]
  \item Eros Quesada \email{eros.quesada@slu.se} [contributor]
  \item Jonathan Stounberg \email{jostou@aqua.dtu.dk} [contributor]
}

}
\keyword{internal}
