% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectanglesToRegions.R
\name{rectanglesToRegions}
\alias{rectanglesToRegions}
\title{rectanglesToRegions}
\usage{
rectanglesToRegions(
  data,
  from = c("ICES Rectangles", "ICES DivAndSubdivision", "ICES Area", "ICES Subarea",
    "ICES Division", "ICES Subdivision", "ICES Regions"),
  to = c("ICES Rectangles", "ICES DivAndSubdivision", "ICES Area", "ICES Subarea",
    "ICES Division", "ICES Subdivision", "ICES Regions"),
  type = c("code", "name"),
  sf = c(TRUE, FALSE),
  plot = c(TRUE, FALSE)
)
}
\arguments{
\item{data}{A vector of input to be transformed}

\item{from}{The spatial resolution of the input. Possible values are "ICES Rectangles", "ICES Subdivision", "ICES Regions".}

\item{to}{The spatial resolution of the output. Possible values are "ICES Rectangles", "ICES Subdivision", "ICES Regions".}

\item{type}{Should the conversion provide the a code (e.g. 27.1.a) or full name (e.g. Barents Sea - NEAFC Regulatory Area; available for all levels, except ICES Rectangle)?}

\item{sf}{Dichotomous, should the result include geometries? Possible values are TRUE or FALSE.}

\item{plot}{Dichotomous, should the result be plotted? Possible values are TRUE or FALSE.}
}
\value{
A vector and a plot, the latter depending on user selection.
}
\description{
rectanglesToRegions
}
\examples{
\dontrun{


}

}
