% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledOrNotsampled.R
\name{sampledOrNotsampled}
\alias{sampledOrNotsampled}
\title{Provides graphical outputs to determine a percentage of sampled data.}
\usage{
sampledOrNotsampled(
  RDBESDataObject,
  year = NA,
  vesselFlag = NA,
  verbose = FALSE
)
}
\arguments{
\item{RDBESDataObject}{An RDBESDataObject.}

\item{year}{Selected Year/Years. The default is NA, i.e. all Years presented in RDBESDataObject.}

\item{vesselFlag}{Selected Country. The default is NA, i.e. all Countries presented in RDBESDataObject.}

\item{verbose}{(Optional) Set to TRUE if you want informative text on
validation and function steps printed out, or FALSE if you don't.
The default is FALSE.}
}
\description{
Provides graphical outputs to determine a percentage of sampled data.
}
