% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shoreline}
\alias{shoreline}
\title{Shoreline spatial data}
\format{
A SpatialPolygonsDataFrame
}
\source{
GSHHG
}
\usage{
shoreline
}
\description{
A data frame containing spatial information for shorelines
}
\keyword{datasets}
