% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageSpatial.R
\name{spatialPlotCreateGraph}
\alias{spatialPlotCreateGraph}
\title{Internal function to create spatial plot using stat rectangles and
(optionallly) circles to represent the sampling coverage}
\usage{
spatialPlotCreateGraph(
  dataToPlot,
  pointsDataToPlot = NA,
  tooltipText = "",
  pointsTooltipText = NA,
  legendText = "",
  pointsLegendText = NA,
  title = "",
  subtitle = ""
)
}
\arguments{
\item{dataToPlot}{Data to plot by stat rectangle}

\item{pointsDataToPlot}{Optional secondary data to plot as circles}

\item{tooltipText}{Tooltip text}

\item{pointsTooltipText}{Tooltip text for secondary data}

\item{legendText}{Legend text}

\item{pointsLegendText}{Legend text for secondary data}

\item{title}{Title}

\item{subtitle}{Subtitle}
}
\value{
A ggplot graph
}
\description{
Internal function to create spatial plot using stat rectangles and
(optionallly) circles to represent the sampling coverage
}
