% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{speciesNamesAndCodes}
\alias{speciesNamesAndCodes}
\title{Species names and codes.}
\format{
A data frame containing the species information
\describe{
  \item{FAOCode}{Three letter FAO code for species}
  \item{FAODescription}{Text description of the FAO code}
  \item{ScientificName}{Scientific name of the species}
  \item{AphiaID}{Aphia ID of the species}
}
}
\source{
David Currie
}
\usage{
speciesNamesAndCodes
}
\description{
A data frame containing the FAO code, FAO description, Aphia ID, and
scientific name for a selection of species
}
\keyword{datasets}
