% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{temporalPlot}
\alias{temporalPlot}
\title{Internal function to return a list of plots which compare the relative
amount of the values of landingVariable, effortVariable and samplingVariable
by quarter.}
\usage{
temporalPlot(
  landingsData = NA,
  effortData = NA,
  sampleData = NA,
  vesselFlag,
  catchCat,
  landingsVariable,
  effortVariable,
  samplingVariable
)
}
\arguments{
\item{landingsData}{Landings data}

\item{sampleData}{Sample data}

\item{vesselFlag}{Registered Country of Vessel - e.g "IE", "ES" or "FR".}

\item{catchCat}{Catch category}

\item{samplingVariable}{variable from SA to plot}

\item{commercialVariable}{Variable from CL to plot}
}
\value{
A tagList of plotly plots
}
\description{
Internal function to return a list of plots which compare the relative
amount of the values of landingVariable, effortVariable and samplingVariable
by quarter.
}
