% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validateCoverageParameters}
\alias{validateCoverageParameters}
\title{Interntal function to validate the values of the parameters used in the
coveragexxx functions()}
\usage{
validateCoverageParameters(
  year,
  quarter,
  vesselFlag,
  landingsVariable,
  effortVariable,
  samplingVariable,
  groupingVariable,
  catchCat,
  includeLandings,
  includeEffort,
  includeSamples,
  topN
)
}
\arguments{
\item{year}{Parameter to validate}

\item{quarter}{Parameter to validate}

\item{vesselFlag}{Parameter to validate}

\item{landingsVariable}{Parameter to validate}

\item{effortVariable}{Parameter to validate}

\item{samplingVariable}{Parameter to validate}

\item{groupingVariable}{Parameter to validate}

\item{catchCat}{Parameter to validate}

\item{includeLandings}{Parameter to validate}

\item{includeEffort}{Parameter to validate}

\item{includeSamples}{Parameter to validate}

\item{topN}{Parameter to validate}
}
\description{
Interntal function to validate the values of the parameters used in the
coveragexxx functions()
}
