% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wormsSpecies}
\alias{wormsSpecies}
\title{Species WORMS aphiaids and descriptions.}
\format{
A data frame containing the species information
\describe{
  \item{Key}{WORMS aphiaid}
  \item{Description}{Species name}
}
}
\source{
ICES
}
\usage{
wormsSpecies
}
\description{
A data frame containing the WORMS aphiaid, and
scientific name for a selection of species.  See
https://www.marinespecies.org/
}
\keyword{datasets}
