% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmountByGroup.R
\name{amountByGroup}
\alias{amountByGroup}
\title{Provides barplots of the inputted table column, summed by up til 4 factors,
inserted as the column names from the same table.}
\usage{
amountByGroup(
  RDBESobj = myH1RawObject,
  var = "CEnumDomTrip",
  valBy = c("CEmetier6"),
  output_type = "plot"
)
}
\arguments{
\item{RDBESobj}{an RDBESDataObject)}

\item{var}{one column name from the RDBESDataObject)}

\item{valBy}{one or more column name from the same table as the 'var' variable)
The order of the column names put in the 'valBy' determines how the plot is arranged
columns with many factorlevels can be put in the end to get a new plot pr level.}

\item{output_type}{Should the output be a bar plot or a table, written to the environment)}
}
\description{
Provides barplots of the inputted table column, summed by up til 4 factors,
inserted as the column names from the same table.
}
