% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageLandings.R
\name{bivariatePlot}
\alias{bivariatePlot}
\title{Internal function to return a list of plots which compare the statistical
rectangle where landings occured to the statistical rectangles where
sampling occured.  The relative amounts of landings/samples are
shown by the use of different colours for the rectangle.}
\usage{
bivariatePlot(
  landingsData,
  sampleData,
  flagLabel,
  catchCat,
  commercialVariable,
  samplingVariable
)
}
\arguments{
\item{landingsData}{Landings data}

\item{sampleData}{Sample data}

\item{flagLabel}{Text to use for vessel flag description}

\item{catchCat}{Catch category}

\item{commercialVariable}{The variable from the landings data to plot}

\item{samplingVariable}{The variable from the sample data to plot}
}
\value{
A tagList of ggplot2 plots
}
\description{
Internal function to return a list of plots which compare the statistical
rectangle where landings occured to the statistical rectangles where
sampling occured.  The relative amounts of landings/samples are
shown by the use of different colours for the rectangle.
}
