% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createBarPlot}
\alias{createBarPlot}
\title{Internal function to create a bar plot}
\usage{
createBarPlot(dataToPlot, formulaForGraph, title)
}
\arguments{
\item{dataToPlot}{The data to plot}

\item{formulaForGraph}{The formula for the x axis value}

\item{title}{Title to use for the plot}
}
\value{
A plotly plot
}
\description{
Internal function to create a bar plot
}
