% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterEffortDataForCoverage}
\alias{filterEffortDataForCoverage}
\title{Internal function to filter effort data for coverageXXX functions}
\usage{
filterEffortDataForCoverage(effortData, year, quarter, vesselFlag, verbose)
}
\arguments{
\item{effortData}{A data table of effort data ("CE")}

\item{year}{Year to be assessed e.g 2021}

\item{quarter}{Quarter to be assessed - possible choices 1,2,3 or 4}

\item{vesselFlag}{Registered Country of Vessel - e.g "IE", "ES" or "FR"}

\item{verbose}{(Optional) Set to TRUE to print more information. Default is
FALSE}
}
\value{
A data table of effort data ("CE")
}
\description{
Internal function to filter effort data for coverageXXX functions
}
