% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterSampleDataForCoverage}
\alias{filterSampleDataForCoverage}
\title{Internal function to filter sample data for coverageXXX functions}
\usage{
filterSampleDataForCoverage(
  sampleData,
  year,
  quarter,
  vesselFlag,
  catchCat,
  verbose
)
}
\arguments{
\item{sampleData}{An RDBESEstObject}

\item{year}{Year to be assessed e.g 2021}

\item{quarter}{Quarter to be assessed - possible choices 1,2,3 or 4}

\item{vesselFlag}{Registered Country of Vessel - e.g "IE", "ES" or "FR".}

\item{catchCat}{Sampling catch category - landings, catch or discards}

\item{verbose}{(Optional) Set to TRUE to print more information. Default is
FALSE}
}
\value{
An RDBESEstObject
}
\description{
Internal function to filter sample data for coverageXXX functions
}
