% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageByGear.R
\name{gearPlot}
\alias{gearPlot}
\title{Internal function to return a list of plots which compare the fishing gears
in landings, effort, and sample data.}
\usage{
gearPlot(
  landingsData = NA,
  effortData = NA,
  sampleData = NA,
  vesselFlag,
  catchCat,
  quarter,
  landingsVariable,
  effortVariable,
  samplingVariable
)
}
\arguments{
\item{landingsData}{Landings data}

\item{effortData}{Effort data}

\item{sampleData}{Sample data}

\item{vesselFlag}{Registered Country of Vessel - e.g "IE", "ES" or "FR".}

\item{catchCat}{Catch category}

\item{quarter}{Quarter of year}

\item{landingsVariable}{Variable from CE to plot}

\item{samplingVariable}{variable from SA to plot}
}
\value{
A tagList of plotly plots
}
\description{
Internal function to return a list of plots which compare the fishing gears
in landings, effort, and sample data.
}
