% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pointsPlot}
\alias{pointsPlot}
\title{Internal function to return a list of plots which compare the statistical
rectangle where landings occured to the statistical rectangles where
sampling occured. The relative amounts of landings/samples are
shown by i) coloured rectangles for landings, and ii) circles for
sampling.}
\usage{
pointsPlot(
  landingsData,
  sampleData,
  vesselFlag,
  catchCat,
  commercialVariable,
  samplingVariable
)
}
\arguments{
\item{landingsData}{Landings data}

\item{sampleData}{Sample data}

\item{vesselFlag}{Registered Country of Vessel - e.g "IE", "ES" or "FR".}

\item{catchCat}{Catch category}

\item{commercialVariable}{The variable from the landings data to plot}

\item{samplingVariable}{The variable from the sample data to plot}
}
\value{
A tagList of ggplot2 plots
}
\description{
Internal function to return a list of plots which compare the statistical
rectangle where landings occured to the statistical rectangles where
sampling occured. The relative amounts of landings/samples are
shown by i) coloured rectangles for landings, and ii) circles for
sampling.
}
