% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageBySpecies.R
\name{speciesPlot}
\alias{speciesPlot}
\title{Internal function to return a list of plots which compare the species in
landings to those in the sample data.}
\usage{
speciesPlot(
  landingsData = NA,
  sampleData = NA,
  vesselFlag,
  catchCat,
  topN = 10
)
}
\arguments{
\item{landingsData}{Landings data}

\item{sampleData}{Sample data}

\item{vesselFlag}{Registered Country of Vessel - e.g "IE", "ES" or "FR".}

\item{catchCat}{Catch category}

\item{(Optional)}{Number of species to plot.  Default is 10.}
}
\value{
A tagList of plotly plots
}
\description{
Internal function to return a list of plots which compare the species in
landings to those in the sample data.
}
