% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{temporalPlot}
\alias{temporalPlot}
\title{Internal function to return a list of plots which compare the relative
amount of the values of commercialVariable and samplingVariable by quarter
landings to those in the sample data.}
\usage{
temporalPlot(
  landingsData,
  sampleData,
  vesselFlag,
  catchCat,
  commercialVariable,
  samplingVariable
)
}
\arguments{
\item{landingsData}{Landings data}

\item{sampleData}{Sample data}

\item{vesselFlag}{Registered Country of Vessel - e.g "IE", "ES" or "FR".}

\item{catchCat}{Catch category}

\item{commercialVariable}{Variable from CL to plot}

\item{samplingVariable}{variable from SA to plot}
}
\value{
A tagList of plotly plots
}
\description{
Internal function to return a list of plots which compare the relative
amount of the values of commercialVariable and samplingVariable by quarter
landings to those in the sample data.
}
